+++ workspace.go
@@ -122,6 +122,7 @@ type Workspace struct {
 	AllowDestroyPlan           bool                  `jsonapi:"attr,allow-destroy-plan"`
 	AssessmentsEnabled         bool                  `jsonapi:"attr,assessments-enabled"`
 	AutoApply                  bool                  `jsonapi:"attr,auto-apply"`
+	AutoDestroyAt              *time.Time            `jsonapi:"attr,auto-destroy-at,iso8601"`
 	CanQueueDestroyPlan        bool                  `jsonapi:"attr,can-queue-destroy-plan"`
 	CreatedAt                  time.Time             `jsonapi:"attr,created-at,iso8601"`
 	Description                string                `jsonapi:"attr,description"`
@@ -296,6 +297,9 @@ type WorkspaceCreateOptions struct {
 	// Optional: Whether to automatically apply changes when a Terraform plan is successful.
 	AutoApply *bool `jsonapi:"attr,auto-apply,omitempty"`
 
+	// Optional: The time after which an automatic destroy run will be queued
+	AutoDestroyAt *time.Time `jsonapi:"attr,auto-destroy-at,iso8601"`
+
 	// Optional: A description for the workspace.
 	Description *string `jsonapi:"attr,description,omitempty"`
 
@@ -418,6 +422,9 @@ type WorkspaceUpdateOptions struct {
 
 	// Optional: Whether to automatically apply changes when a Terraform plan is successful.
 	AutoApply *bool `jsonapi:"attr,auto-apply,omitempty"`
+
+	// Optional: The time after which an automatic destroy run will be queued
+	AutoDestroyAt *time.Time `jsonapi:"attr,auto-destroy-at,iso8601"`
 
 	// Optional: A new name for the workspace, which can only include letters, numbers, -,
 	// and _. This will be used as an identifier and must be unique in the